Move RNG to a separate file
