+++ torchquad/integration/vegas.py
@@ -4,7 +4,8 @@ from loguru import logger
 
 
 from .base_integrator import BaseIntegrator
-from .utils import _setup_integration_domain, RNG
+from .utils import _setup_integration_domain
+from .rng import RNG
 from .vegas_map import VEGASMap
 from .vegas_stratification import VEGASStratification
