+++ torchquad/tests/utils_integration_test.py
@@ -3,7 +3,7 @@ import sys
 sys.path.append("../")
 
 from autoray import numpy as anp
-from autoray import infer_backend, get_dtype_name, to_backend_dtype, to_numpy
+from autoray import infer_backend, get_dtype_name, to_backend_dtype
 import importlib
 import pytest
 import warnings
@@ -12,7 +12,6 @@ from integration.utils import (
     _linspace_with_grads,
     _add_at_indices,
     _setup_integration_domain,
-    RNG,
 )
 from utils.set_precision import set_precision
 from utils.enable_cuda import enable_cuda
@@ -172,52 +171,11 @@ def test_setup_integration_domain():
     _run_tests_with_all_backends(_run_setup_integration_domain_tests)
 
 
-def _run_RNG_tests(dtype_name, backend):
-    """
-    Test the random number generator with the given dtype and numerical backend
-    * With the same seed, the same numbers should be generated
-    * With different seeds, different numbers should be generated
-    * If seed is None / omitted, the RNG should be randomly seeded
-    """
-    backend_dtype = to_backend_dtype(dtype_name, like=backend)
-    size = [3, 9]
-    generateds = [
-        RNG(backend, 547).uniform(size=size, dtype=backend_dtype),
-        RNG(backend, None).uniform(size=size, dtype=backend_dtype),
-        RNG(backend, 547).uniform(size=size, dtype=backend_dtype),
-        RNG(backend).uniform(size=size, dtype=backend_dtype),
-        RNG(backend, 42).uniform(size=size, dtype=backend_dtype),
-    ]
-    numpy_arrs = list(map(to_numpy, generateds))
-
-    # Validity of the backend, dtype, shape and values range
-    assert all(infer_backend(arr) == backend for arr in generateds)
-    assert all(get_dtype_name(arr) == dtype_name for arr in generateds)
-    assert all(arr.shape == (3, 9) for arr in generateds)
-    assert all(0.0 <= x <= 1.0 for arr in numpy_arrs for x in arr.ravel())
-
-    # Test if the seed argument leads to consistent results and
-    # if omitting a seed leads to random numbers
-    assert anp.array_equal(numpy_arrs[0], numpy_arrs[2])
-    for i1 in range(len(generateds)):
-        for i2 in range(i1 + 1, len(generateds)):
-            if i1 == 0 and i2 == 2:
-                continue
-            # With a very low probability this may fail
-            assert not anp.array_equal(numpy_arrs[i1], numpy_arrs[i2])
-
-
-def test_RNG():
-    """Test the random number generator with all possible configurations"""
-    _run_tests_with_all_backends(_run_RNG_tests)
-
-
 if __name__ == "__main__":
     try:
         # used to run this test individually
         test_linspace_with_grads()
         test_add_at_indices()
         test_setup_integration_domain()
-        test_RNG()
     except KeyboardInterrupt:
         pass
