+++ torchquad/tests/vegas_stratification_test.py
@@ -5,7 +5,7 @@ sys.path.append("../")
 from autoray import numpy as anp
 from autoray import to_backend_dtype
 
-from integration.utils import RNG
+from integration.rng import RNG
 from integration.vegas_stratification import VEGASStratification
 
 from helper_functions import setup_test_for_backend
